refactor: create src directory
